Big Format Update
Lines 25 - 28 -> decided that formatting groups would be easier with a JSON
Line 77 -> added a semicolon, line 68 had one and I wasn't sure which was correct - rlly great job on making the loops make sense btw! i had a big smooth brain there.
Line 103 - 112 -> I have no idea if this will actually work. I'm not hopeful, like, I think it's correct but it's also 1am for me rn.
Line 163 - 194 -> the ramblings of a madman - however it's a LOT closer than I thought it was when I did a bug test (a lot of it was guesswork lol)(normal legends aren't compatible with network graphs (i think? it didn't put "network graph" in the list of incompatible charts - we'll see though!)).